WIP: Ignore: Sketching a new visitor-pattern
This is a sketch of what an automatically-generated visitor implementation
could look like.  The goal of this is to reduce / remove the amount of
hand-written code necessary to make the likes of `tree.Expr` or
`tree.Statement` visitable.

Because we can generate code, this API is aggressively typed. Instead of
requiring a generic Pre()/Post() API, an interface and a composable base type
are generated with a pre/post pair per type of visitable object.

This API can support visitable objects using pointer receivers for in-place
modification as well as value recivers to support structural modification of
immutable-style objects. This is accomplished by providing each visitation
method with a Context that will effect structural modifications on behalf of
the caller.

It's worth noting that a no-op visitor has only amortized allocation cost over
an arbitrarily-sized graph:

```
BenchmarkNoop-8   	 2000000	       653 ns/op	       0 B/op	       0 allocs/op
```

Release note: None